move JobBackend & utility functions to easybuild.tools.job.backend module
